Fix onboarding animation
